Updated description with malaraiEquilibriumVivax. Removed duplicate tests and renamed vivax test files.
Rewrote parts of calculate_infections() to reduce use of to_vector(). It is now only used twice for vivax: for the source vector and for the bitten vector. I can't think of a way around using it twice there. While I was doing this I compared it with the falciparum code, and I don't think to_vector isn't needed there, so I've removed it hoping that it could increase speed there.