Allow simulation to be restored with new interventions.
This uses a few of individual's new (or upcoming) features that make
restoring more flexible. It also fixes a bug when restoring the
mosquitto solvers, by correctly restoring the timestep, which is needed
to model seasonality.

It adds end-to-end testing of this feature, across a range of scenarios.
For each scenario, the outcomes of the simulation with and without
restoring are compared and we make sure they are equivalent.