Allow simulation to be restored with new interventions. (#286)
This uses a few of individual's new features that make restoring more
flexible. It also fixes a bug when restoring the mosquitto solvers, by
correctly restoring the timestep, which is needed to model seasonality.

The intervention events use the new `restore = FALSE` flag to make sure
their schedule can be modified when resuming. Instead of having the
events re-schedule themselves everytime they fire, we setup the entire
schedule upfront when initialising the simulation.

It adds end-to-end testing of this feature, across a range of scenarios.
For each scenario, the outcomes of the simulation with and without
restoring are compared and we make sure they are equivalent.