Rewrite the exponential decay process in C++. (#285)
The existing process written in R needs to copy the contents of each
variable from C++ using `v$get_values()`, then after scaling the vector
it would copy the result back into C++ using `v$queue_update`.

The amount of data copied and the time it took was pretty significant.
6 double variables, one for each kind of immunity, need to be updated in
full at each time step, each as big as the population size.

Moving this into C++ removes the need for any copy at all, besides the
multication loop. Values are read out of a reference to the vector held
by the DoubleVariable, the result of the multication is moved to the
queue, and finally individual moves the vector in the queue into the
DoubleVariable. The speedup from this change for a 1M population size is
around 10%.

An alternative optimization I considered was to compute the exponential
decay lazily, recording only the timestep and value at which the
immunity was last updated and using the closed form expression of the
exponential decay.  This would avoid the need to have mass updates of
the immunity variables at every time step. Unfortunately in my testing
this ends up being slower than even the current implementation, with all
the time being spent in calculating the current value. This would also
be a much more intrusive change, since every use of the immunity
variables needs to be modified to take the last update timestep, the
current timestep and the decay rate into consideration.