Update description of the dt parameter.
The dt parameter was described as "the delay to move from state Tr to
Ph", but the Ph (prophylaxis) state does not exist in the model. Instead
the dt parameter is used to transition from Tr to S, in `R/processes.R`.

Looking at the commit history, there seems to have been a bit of back
and forth around adding and removing the Tr and Ph states. The comment
in question was introduced in mrc-ide/malariasimulation#30, after the
prophylaxis state had already been removed and even at the time the
parameter was already used as the transition rate from Tr to S.