Antiparasite immunity function shapes the duration of subpatent infection/rate of recovery from subpatent infection in the Pv model. These values are now assigned during the variable creation and calculation of disease progression rates. Antiparasite immunity is labelled IAA and IAM for acquired and maternal immunities, where their decay, boosting, maternal immunity and reset at death are now included. The decay of the immunity occurs daily and has an impact of disease progression rate, which is now calculated and updated daily.
We must also adjust the prevalence rendering, which for p.f is based on detectability immunity, and for p.v is explicit in the human states.

`du` and ID variables and parameters are removed for vivax.

Tests are added for du rates and for prevalence rendering.

The vignette is extended to include the removal of ID and the sub-patent duration calculation.