Update vector_parameters.R references
Updated references for Q0 for An. arabiensis, An. gambiae and An. funestus. These estimates are generated by Griffin et al (2010) using data from a number of studies. Previously, it appeared as though all bionomics parameters were obtained from Sherrard-Smith et al (2019) work in PNAS, but this study gives estimates for phi_bednets and phi_indoors for these three vector species. Also clarified that the phi parameters for An. stephensi are for phi_bednets and phi_indoors.